- fixed bug: TileTheme gets saved properly after getting filled from asset file - fixed bug: infinite loop in filling the configurations for 3d themes - base pieces can be mirror on the x axis when checking for matchin transformation - 3D Mesher seems to work, needs testing

+++ Assets/tile_mesher/TileTheme.cs
@@ -105,9 +105,9 @@ namespace MeshBuilder
         private void FillConfigurations()
         {
             List<int> nullConfigs = new List<int>();
-            for (byte i = 0; i < configs.Length; ++i)
+            for (int i = 0; i < configs.Length; ++i)
             {
-                configs[i] = TileConfigurationBuilder.Decompose(i, baseVariants);
+                configs[i] = TileConfigurationBuilder.Decompose((byte)i, baseVariants);
                 if (i > 0 && configs[i].Count == 0)
                 {
                     nullConfigs.Add(i);
